C++ Primer Plus Learning
finish learning chapter 11
To page 424

Time:2018/7/16
Location:home

+++ C++ Code/C++ Primer Plus/Chapter 11/mytime3.cpp
@@ -53,7 +53,7 @@ Time Time::operator-(const Time & t) const
 Time Time::operator*(double mult) const
 {
     Time result;
-    long totalminutes = hours * mult * 60 + minutes * mult;
+    int totalminutes = hours * mult * 60 + minutes * mult;
     result.hours = totalminutes / 60;
     result.minutes = totalminutes % 60;
     return result;
